datastore cli

+++ src/pfcli/credential.py
@@ -95,9 +95,21 @@ def delete(
         ...,
         '--cred-id',
         '-i',
-        help='UUID of credential to delete.'
+        help='UUID of credential to delete.',
+        confirmation_prompt=True,
+    ),
+    force: bool = typer.Option(
+        False,
+        '--force',
+        '-f',
+        help="Forcefully delete credential without confirmation prompt."
     )
 ):
+    if not force:
+        do_delete = typer.confirm("Are your sure to delete credential?")
+        if not do_delete:
+            raise typer.Abort()
+
     client: CredentialClientService = build_client(ServiceType.CREDENTIAL)
     client.delete_credential(cred_id)
